add single node pruning

+++ pruning-scipts-caffe/base_var_gradient.sh
@@ -16,7 +16,8 @@ skip_input_channels=false
 skip_output_channels=false
 retrain=false
 characterise=false
-while getopts ":fircoqp" arg; do
+channel_var=true
+while getopts ":frcv" arg; do
   case $arg in
     c ) # Display help.
       characterise=true
@@ -25,21 +26,12 @@ while getopts ":fircoqp" arg; do
     f ) # Display help.
       force=true
       ;;
-    i ) # Display help.
-      input_channels=true
-      ;;
-    o ) # Display help.
-      output_channels=true
-      ;;
-    q ) # Display help.
-      skip_input_channels=true
-      ;;
-    p ) # Display help.
-      skip_output_channels=true
-      ;;
     r ) # Display help.
       retrain=true
       characterise=false
+      ;;
+    v ) # Display help.
+      channel_var=true
       ;;
   esac
 done
@@ -53,18 +45,12 @@ then
 else
   filename_prefix=$filename_prefix\sensitivity_
 fi
-if [[ $input_channels == true ]]
+
+if [[ $channel_var == true ]]
 then
-  if [[ $output_channels == true ]]
-  then
-    filename_prefix=$filename_prefix\input_output_channels_
-  else
-    filename_prefix=$filename_prefix\input_channels_
-  fi
+  filename_prefix=$filename_prefix\channel_avg_
 fi
-
 saliency_scale=none_scale
-
 mkdir -p $default_save_path
 for (( i=1; i<=$iterations; i++ ))
   do
@@ -96,6 +82,62 @@ for (( i=1; i<=$iterations; i++ ))
       \--test-size $test_size \
       \--eval-size $eval_size \
       \--train-size $train_size \
+      \--channel-var $channel_var \
+      \--scaling $scaling 
+      rm $filename_partial
+    fi
+  fi
+done
+
+channel_var=false
+filename_prefix=$default_save_path/summary_
+if [[ $characterise == true ]]
+then
+  filename_prefix=$filename_prefix\characterise_
+elif [[ $retrain == true ]]
+then
+  filename_prefix=$filename_prefix\retrain_
+else
+  filename_prefix=$filename_prefix\sensitivity_
+fi
+
+if [[ $channel_var == true ]]
+then
+  filename_prefix=$filename_prefix\channel_avg_
+fi
+saliency_scale=none_scale
+mkdir -p $default_save_path
+for (( i=1; i<=$iterations; i++ ))
+  do
+  if [[ $i == 1 ]] && [[ $characterise == false ]] && [[ $retrain == false ]]
+  then
+    use_stop_acc=10.0
+  else
+    use_stop_acc=$stop_acc
+  fi
+  input_lower=${input,,}
+  saliency_method_lower=${saliency_method,,}
+  norm_lower=${norm,,}
+  filename=$filename_prefix\_activation_var_gradient_mean-none_norm-no_normalisation\_python_iter$i.npy
+  if [[ ! -e $filename ]] || [[ $force == true ]]
+  then
+    filename_partial=$filename\.partial
+    if [[ ! -e $filename_partial ]]
+    then
+      touch $filename_partial
+      echo $filename
+      GLOG_minloglevel=1 python compare_pruning_techniques_variance.py \
+      \--filename $filename \
+      \--arch $arch \
+      \--dataset $dataset \
+      \--stop-acc $use_stop_acc \
+      \--characterise $characterise \
+      \--retrain $retrain \
+      \--test-interval $test_interval \
+      \--test-size $test_size \
+      \--eval-size $eval_size \
+      \--train-size $train_size \
+      \--channel-var $channel_var \
       \--scaling $scaling 
       rm $filename_partial
     fi
